the problem of working with negative numbers and decimal fractions has been solved

+++ libs/fsyscal.h
@@ -4,6 +4,6 @@
 #define MAX 10000
 
 void operator_input(void);
-void operand_input(long int *p);
+void operand_input(double *p);
 
 #endif
